Basic test of refacto
